pyproject.toml
[project]
name = "auth_project"
version = "0.1.0"
description = ""
authors = [
    {name = "ayamushev",email = "ayamushev@rvision.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all] (>=0.117.1,<0.118.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "psycopg (>=3.2.10,<4.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "psycopg-binary (>=3.2.10,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "cachetools (>=6.2.0,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
